@@ -1,15 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Equinox.CosmosStore" Version="3.0.7" />
-    <PackageReference Include="Equinox.CosmosStore.Prometheus" Version="3.0.7" />
-    <PackageReference Include="Equinox.EventStore" Version="3.0.7" />
-    <PackageReference Include="Equinox.MemoryStore" Version="3.0.7" />
-    <PackageReference Include="FsCodec.SystemTextJson" Version="2.3.2" />
+<!--    <PackageReference Include="Equinox.CosmosStore" Version="4.0.0-beta.13" />-->
+<!--    <PackageReference Include="Equinox.CosmosStore.Prometheus" Version="4.0.0-beta.13" />-->
+<!--    <PackageReference Include="Equinox.EventStore" Version="4.0.0-beta.13" />-->
+<!--    <PackageReference Include="Equinox.MemoryStore" Version="4.0.0-beta.13" />-->
     <PackageReference Include="prometheus-net.AspNetCore" Version="3.6.0" />
     <PackageReference Include="Serilog.AspNetCore" Version="3.2.0" />
     <PackageReference Include="Serilog.Sinks.Console" Version="4.0.0" />
