@@ -1,12 +1,7 @@
-﻿using Equinox;
-using Equinox.Core;
-using Microsoft.FSharp.Core;
-using Newtonsoft.Json;
-using Serilog;
+﻿using Microsoft.FSharp.Core;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using FsCodec;
 
 namespace TodoBackendTemplate
 {
@@ -21,45 +16,37 @@ namespace TodoBackendTemplate
 
             public class Snapshotted : Event
             {
-                public new bool Happened { get; set; }
+                public bool HasHappened { get; set; }
             }
 
-            static readonly JsonNetUtf8Codec Codec = new JsonNetUtf8Codec(new JsonSerializerSettings());
+            static readonly SystemTextJsonUtf8Codec Codec = new(new());
             
-            public static Event TryDecode(string et, byte[] json)
-            {
-                switch (et)
+            public static FSharpValueOption<Event> TryDecode(string et, byte[] json) =>
+                et switch
                 {
-                    case nameof(Happened): return Codec.Decode<Happened>(json);
-                    case nameof(Snapshotted): return Codec.Decode<Snapshotted>(json);
-                    default: return null;
-                }
-            }
+                    nameof(Happened) => Codec.Decode<Happened>(json),
+                    nameof(Snapshotted) => Codec.Decode<Snapshotted>(json),
+                    _ => FSharpValueOption<Event>.None
+                };
 
-            public static Tuple<string, byte[]> Encode(Event e) => Tuple.Create(e.GetType().Name, Codec.Encode(e));
-            public static string For(ClientId id) => StreamNameModule.create("Aggregate", id.ToString());
+            public static (string, ReadOnlyMemory<byte>) Encode(Event e) => (e.GetType().Name, Codec.Encode(e));
+            public static (string, string) StreamIds(ClientId id) => ("Aggregate", id.ToString());
         }
         public class State
         {
             public bool Happened { get; set; }
 
-            internal State(bool happened) { Happened = happened; }
+            State(bool happened) { Happened = happened; }
 
-            public static readonly State Initial = new State(false);
+            public static readonly State Initial = new (false);
 
-            static void Evolve(State s, Event x)
-            {
-                switch (x)
+            static void Evolve(State s, Event x) =>
+                s.Happened = x switch
                 {
-                    case Event.Happened e:
-                        s.Happened = true;
-                        break;
-                    case Event.Snapshotted e:
-                        s.Happened = e.Happened;
-                        break;
-                    default: throw new ArgumentOutOfRangeException(nameof(x), x, "invalid");
-                }
-            }
+                    Event.Happened e => true,
+                    Event.Snapshotted e => e.HasHappened,
+                    _ => throw new ArgumentOutOfRangeException(nameof(x), x, "invalid")
+                };
 
             public static State Fold(State origin, IEnumerable<Event> xs)
             {
@@ -72,7 +59,7 @@ namespace TodoBackendTemplate
 
             public static bool IsOrigin(Event e) => e is Event.Snapshotted;
             
-            public static Event Snapshot(State s) => new Event.Snapshotted {Happened = s.Happened};
+            public static Event Snapshot(State s) => new Event.Snapshotted {HasHappened = s.Happened};
         }
 
         /// Defines the decision process which maps from the intent of the `Command` to the `Event`s that represent that decision in the Stream 
@@ -82,56 +69,39 @@ namespace TodoBackendTemplate
             {
             }
 
-            public static IEnumerable<Event> Interpret(State s, Command x)
+            public IEnumerable<Event> Interpret(State s)
             {
-                switch (x)
+                switch (this)
                 {
                     case MakeItSo c:
                         if (!s.Happened) yield return new Event.Happened();
                         break;
-                    default: throw new ArgumentOutOfRangeException(nameof(x), x, "invalid");
+                    default: throw new ArgumentOutOfRangeException("this", this, "invalid");
                 }
             }
         }
 
-        class Handler
-        {
-            readonly EquinoxStream<Event, State> _stream;
-
-            public Handler(ILogger log, IStream<Event, State> stream) =>
-                _stream = new EquinoxStream<Event, State>(State.Fold, log, stream);
-
-            /// Execute `command`, syncing any events decided upon
-            public Task<Unit> Execute(Command c) =>
-                _stream.Execute(s => Command.Interpret(s, c));
-
-            /// Establish the present state of the Stream, project from that as specified by `projection`
-            public Task<T> Query<T>(Func<State, T> projection) =>
-                _stream.Query(projection);
-        }
-
-        public class View
-        {
-            public bool Sorted { get; set; }
-        }
+        public record View(bool Sorted);
 
         public class Service
         {
             /// Maps a ClientId to Handler for the relevant stream
-            readonly Func<ClientId, Handler> _stream;
+            readonly Func<ClientId, Equinox.DeciderCore<Event, State>> _resolve;
 
-            public Service(ILogger handlerLog, Func<string, IStream<Event, State>> resolve) =>
-                _stream = id => new Handler(handlerLog, resolve(Event.For(id)));
+            public Service(Func<ClientId, Equinox.DeciderCore<Event, State>> resolve) =>
+                _resolve = resolve; 
 
             /// Execute the specified command 
             public Task<Unit> Execute(ClientId id, Command command) =>
-                _stream(id).Execute(command);
+                _resolve(id).Transact(command.Interpret);
 
             /// Read the present state
             // TOCONSIDER: you should probably be separating this out per CQRS and reading from a denormalized/cached set of projections
-            public Task<View> Read(ClientId id) => _stream(id).Query(Render);
+            public Task<View> Read(ClientId id) =>
+                _resolve(id).Query(Render);
 
-            static View Render(State s) => new View() {Sorted = s.Happened};
+            static View Render(State s) =>
+                new (Sorted: s.Happened);
         }
     }
 }
